Module Lib.BSTSet

type 'a tree =

| Empty

| Node of 'a tree * 'a * 'a tree

Datatype for polymorphic binary trees

val tree_of_sexp : 'a. ( Sexplib0.Sexp.t -> 'a ) -> Sexplib0.Sexp.t -> 'a tree

val sexp_of_tree : 'a. ( 'a -> Sexplib0.Sexp.t ) -> 'a tree -> Sexplib0.Sexp.t

val leaf : 'a -> 'a tree

val lookup : 'a -> 'a tree -> bool

lookup x t searches the BST t for the element x & returns a boolean lookup leverages the BST invariants so we don't have to search every subtree

val tree_max : 'a tree -> 'a

tree_max t finds the max element in a binary tree, using the polymorphic max function

val inorder : 'a tree -> 'a list

inorder t performs an inorder traversal of the BST t

val print_tree : int tree -> unit

Prints an int tree to the console

module BSTSet : SetInterface.SetInterface