Module Lib.CmdLineParser

Code adapted from Real World OCaml, Chapter 16

val prompt_for_string : string -> ( string -> 'a ) -> 'a

Prompts the user to enter a value, where name is the name of the value, and of_string if a function converting the string input to type 'a

val anon_prompt : string -> ( string -> 'a ) -> 'a Core.Command.Param.t

Creates a parser that automatically prompts the user with the message name if a value isn’t provided, where of_string is used to convert the user input from string to type 'a

val is_filename : string -> string

Checks if a string is a valid readable filename

val regular_file : string Core.Command.Arg_type.t

Defines a cmd-line argument type regular_file which ensures that the input file isn't some Unix file type that can’t be read

val anon_prompt_reg_file : string -> string Core.Command.Param.t

TODO: not sure why this doesn't work