Module Lib.Parser

Parser utility functions, based on the Angstrom parser-combinator library.

The implementation of some parsers is adapted from the Parsers lecture of Penn's CIS 5520 Haskell course.

module A = Angstrom

Alias for the Angstrom module

Boolean predicates on characters

val is_digit : Base.char -> Base.bool

is_digit c returns true if c is a digit, false otherwise

val is_uppercase : Base.char -> Base.bool

is_uppercase c returns true if c is a capital letter, false otherwise

val is_lowercase : Base.char -> Base.bool

is_lowercase c returns true if c is a lowercase letter, false otherwise

val is_letter : Base.char -> Base.bool

is_letter c returns true if c is a letter, false otherwise

val is_whitespace : Base.char -> Base.bool

is_whitespace c returns true if c is a whitespace character, i.e. tab/newline/space/carriage return

Simple parsers

val lowercaseP : Base.char A.t

Parser for a lowercase letter

val upperCaseP : Base.char A.t

Parser for a capital letter

val letterP : Base.char A.t

Parser for a single letter

val digitP : Base.char A.t

Parser for a single digit

val underscoreP : Base.char A.t

Parser for a single underscore character

val quoteP : Base.char A.t

Parser for a single quote character

val intP : Base.int A.t

Parser for integers

val whitespace : Base.char A.t

Parser for a whitespace character

val wsP : 'a A.t -> 'a A.t

Takes a parser p, runs it & skips over any whitespace chars occurring afterwards

val stringP : Base.string -> Base.unit A.t

Accepts only a particular string str & consumes any white space that follows

val constP : Base.string -> 'a -> 'a A.t

Accepts a particular string s, returns a given value x, and consume any white space that follows

val between : 'b A.t -> 'a A.t -> 'c A.t -> 'a A.t

between openP p closeP parses openP, followed by p and finally closeP. Only the value parsed by p is returned.

val parens : 'a A.t -> 'a A.t

parens p takes a parser p and parses it between parentheses

val chainl1 : 'a A.t -> ( 'a -> 'a -> 'a ) A.t -> 'a A.t

chainl1 e op parses one or more occurrences of e, separated by op, & returns a left-associative application of the functions returned by the op parser onto the values retruned by the e parser.

(Implementation of chainl1 taken from the Angstrom docs)

Functions for running parsers

val run_parser : 'a A.t -> Base.string -> ( 'a, Base.string ) Base.Result.t

run_parser p s uses the parser p to parse the string s