Module Lib.ParserTypes

Datatype definitions for OCaml module signatures. These types are used by the module Lib.ModuleParser to parse OCaml module signatures. This file also calls some helper functions defined in the module Lib.Utils.

type moduleName = Base.string

Name of a module

val moduleName_of_sexp : Sexplib0.Sexp.t -> moduleName

val sexp_of_moduleName : moduleName -> Sexplib0.Sexp.t

type moduleType =

| Intf

| Impl

A module can either be an interface Intf or an implementation Impl

val moduleType_of_sexp : Sexplib0.Sexp.t -> moduleType

val sexp_of_moduleType : moduleType -> Sexplib0.Sexp.t

type ident = Base.string

Identifiers in OCaml

val ident_of_sexp : Sexplib0.Sexp.t -> ident

val sexp_of_ident : ident -> Sexplib0.Sexp.t

type ty =

| Int

| Char

| Bool

| Unit

| Alpha

| AlphaT

| T

| Option of ty

| Pair of ty * ty

| List of ty

| Func1 of ty * ty

| Func2 of ty * ty * ty

ty is an ADT representing concrete types, where:

Alpha represents 'a

AlphaT represents 'a t

T represents t

Option represents option types

Pair represents a pair, i.e. a product type with two arguments

List represents lists

Func1 represents functions of arity 1 (arg type, return type)

Func2 represents functions of arity 2 (arg1 type, arg2 type, return type)

All other constructors correspond to base OCaml types

val ty_of_sexp : Sexplib0.Sexp.t -> ty

val sexp_of_ty : ty -> Sexplib0.Sexp.t

val compare_ty : ty -> ty -> Base.int

val tyEqual : ty -> ty -> Base.bool

Structural equality function for the ty datatype

val string_of_ty : ?alpha:Base.string -> ?t:Base.string -> ?camelCase:bool -> ty -> Base.string

Converts a ty to its string representation, taking in optional arguments specifying various parameters of the string represntation.

Note that AlphaT & T are converted to "expr"

The argument alpha specifies a concrete base type that should be printed in the PBT code in lieu of \'a when testing polymorphic functions.

The argument t specifies a string that should be printed whenever ty is equal to AlphaT or T (eg. the string literal "expr" when generating the expr ADT definition)

The argument camelCase specifies whether the string representation of ty should be in camel-case or not. This is relevant for parameterized types whose string representation may contain a space, eg. "int option".

type abstractType =

| T0

| T1 of ty

Abstract type contained within a module

The T0 constructor means the abstract type is not parameterized by any type variables, i.e. the abstract type is just t

The T1 constructor takes a type ty as its argument, representing an abstract type containing a type variable, eg. 'a t

val abstractType_of_sexp : Sexplib0.Sexp.t -> abstractType

val sexp_of_abstractType : abstractType -> Sexplib0.Sexp.t

type valDecl = {

valName : Base.string;

valType : ty;

}

Type representing a value declaration:

valName is the name of the variable/function being declared

valType is the type of the declaration

E.g. for val empty : 'a t, valName = "empty" and valType = AlphaT

val valDecl_of_sexp : Sexplib0.Sexp.t -> valDecl

val sexp_of_valDecl : valDecl -> Sexplib0.Sexp.t

val valType : valDecl -> ty

val valName : valDecl -> Base.string

module Fields_of_valDecl : sig ... end

type argLabel = valDecl

Type representing the argument to a function

val argLabel_of_sexp : Sexplib0.Sexp.t -> argLabel

val sexp_of_argLabel : argLabel -> Sexplib0.Sexp.t

type moduleSig = {

moduleName : moduleName;

moduleType : moduleType;

abstractType : abstractType;

valDecls : valDecl Base.list;

}

Record type representing an ML module

val moduleSig_of_sexp : Sexplib0.Sexp.t -> moduleSig

val sexp_of_moduleSig : moduleSig -> Sexplib0.Sexp.t

val valDecls : moduleSig -> valDecl Base.list

val abstractType : moduleSig -> abstractType

val moduleType : moduleSig -> moduleType

val moduleName : moduleSig -> moduleName

module Fields_of_moduleSig : sig ... end