Module Lib.Utils

This file contains general utility functions that are called by the other files.

val writeDoc : Stdio.Out_channel.t -> PPrint.document -> Base.unit

writeDoc outc doc writes the PPrint document doc to an Out_channel doc (eg. stdout)

val spaceL : PPrint.document -> PPrint.document

spaceL doc adds a space on the left of the PPrint document doc

val spaceR : PPrint.document -> PPrint.document

spaceR doc adds a space on the right of the PPrint document doc

val spaceLR : PPrint.document -> PPrint.document

spaceLR doc adds a space on both sides of the PPrint document doc

val sBar : PPrint.document

Produces a PPrint document " | " with spaces on both sides

val barSpace : PPrint.document

Produces a PPrint document "| " with a space on the right only

val sArrow : PPrint.document

Produces a PPrint document " -> " with spaces on either side

val star2 : PPrint.document

Produces a PPrint document ** with spaces on either side

val sIn : PPrint.document

Produces a PPrint document " in " with spaces on either side

val comment : PPrint.document -> PPrint.document

Takes a PPrint document body and wraps it in the OCaml comment syntax, i.e. comment body is displayed as (** body *)

val getModuleSigName : Base.string -> Base.string

Given a filepath to a .ml/.mli file, retrieves the corresponding name of the top-level module signature.

Note: The name of the module signature must be the same as the .ml/.mli file.

val replicate : n:Base.int -> 'a -> 'a Base.list

replicate n a produces a list containing n copies of a

val map2 : f:( 'a -> 'b ) -> ('a * 'a) -> 'b * 'b

Applies a function pointwise on a pair

val map3 : f:( 'a -> 'b ) -> ('a * 'a * 'a) -> 'b * 'b * 'b

Applies a function pointwise on a triple

val curry3 : ( ('a * 'b * 'c) -> 'd ) -> 'a -> 'b -> 'c -> 'd

Converts a curried arity-3 function to its uncurried equivalent

val uncurry3 : ( 'a -> 'b -> 'c -> 'd ) -> ('a * 'b * 'c) -> 'd

Converts an uncurried arity-3 function to its curried equivalent