Module Lib.CodeGenerator

Generic utility functions

val write_doc : Stdio.Out_channel.t -> PPrint.document -> Base.unit

Writes a PPrint document to an Out_channel (eg. stdout)

val spaced : PPrint.document -> PPrint.document

spaced doc adds a space on either side of the PPrint document doc

val sBar : PPrint.document

Aliases for PPrint documents for common OCaml symbols

val sArrow : PPrint.document

val star2 : PPrint.document

val comment : PPrint.document -> PPrint.document

Takes a PPrint document body and wraps it in the OCaml comment syntax, i.e. comment body is displayed as (** body *)

val getModuleSigName : Base.string -> Base.string

Given a filepath to a .ml/.mli file, retrieves the corresponding name of the top-level module signature (must be the same as the .ml/.mli file)

val replicate : n:Base.int -> 'a -> 'a Base.list

replicate n a produces a list containing n copies of a

Code generators for Generated.ml, the generated PBT code

val imports : Base.string -> Base.string -> Base.string -> PPrint.document

imports filepath prints out a PPrint document that imports the requisite modules for the PBT code. The sigName, modName1, modName2 arguments are the names of the module signatures & the two module implementations, which must be the same as their corresponding .ml files.

val sexpAnnotation : PPrint.document

Document for printing the PPX annotation for S-Expr serialization (indented), followed by a newline

val isArrowType : ParserTypes.valDecl -> Base.bool

isArrowType v returns true if the value declaration v has an arrow type

val tyIsArrow : ParserTypes.ty -> Base.bool

tyIsArrow ty returns true if ty correspons to an arrow type

val extractArgTypes : ParserTypes.valDecl -> PPrint.document

Extracts the argument types of functions defined in the module signature, and generates constructors for the expr ADT that take these types as type parameters

val exprADTDecl : ParserTypes.moduleSig -> PPrint.document

Generates the definition of the expr ADT

val printConstructor : Base.string -> Base.string Base.list -> PPrint.document

Helper function for printing out OCaml constructors (Wrapper for the OCaml.variant function in the PPrint library)

val ty_of_string : Base.string -> ParserTypes.ty

Converts a string denoting a type to the equivalent ty ADT constructor

val varNameHelper : ParserTypes.ty -> Base.string

varNameHelper ty returns an appropriate variable name corresponding to ty, eg. varNameHelper Int = n

val genVarNamesSingleton : ?prime:bool -> ParserTypes.ty -> Base.string

Special case of genVarNames when we only have one argument type If prime = true, add a single quote to the end of the variable name

val genVarNamesN : n:Base.int -> ParserTypes.ty -> Base.string Base.list

Generates n unique variable names corresponding to a type ty

val genVarNames : ?prime:bool -> ParserTypes.ty Base.list -> Base.string Base.list

Takes a list of argument types, and generates corresponding variable names which are unique for each element of the list eg. genVarNames [Int, Int] = [n1, n2]

val getExprConstructorName : ParserTypes.valDecl -> Base.string

Fetches the expr constructor corresponding to a val declaration in a module

val getExprConstructor : ParserTypes.valDecl -> Base.string Base.list * PPrint.document

Fetches the constructor corresponding to a val declaration in the expr ADT, returning a pair of the form (args, constructor applied to args), eg. (["x", "e"], !^ "Mem(x,e)")

val extractReturnType : ParserTypes.valDecl -> Base.string

Extracts the return type of a function For non-arrow types, this function just extracts the type itself

val uniqRetTypesInSig : ParserTypes.moduleSig -> Base.string Base.list

Fetches the unique return types across the functions / values in a module signature

val tyADTDecl : ParserTypes.moduleSig -> PPrint.document

Generates the definition of the ty ADT

val getFuncName : ParserTypes.valDecl -> PPrint.document

Given a val declaration inside a module (eg. val f : 'a -> 'a), returns the corresponding function name (eg. M.f)

val valADTConstructorString : Base.string -> Base.string

Like valADTConstructor (see below), but returns a string instead of a PPrint document

val valADTConstructor : Base.string -> PPrint.document

valADTConstructor ty generates the constructor name for the value ADT corresponding to the type ty

val tyADTConstructors : ParserTypes.moduleSig -> Base.string Base.list

Generates the list of constructor names for the ty ADT

val tyAndValueADTConstructors : ParserTypes.moduleSig -> (Base.string * Base.string) Base.list

Returns an association list where each element is a constructor for the ty ADT & its associated constructor for the value ADT

val valADTParam : Base.string -> PPrint.document

valADTParam ty generates the type param for the constructor value ADT corresponding to the type ty

val valADTTypeDef : Base.string -> PPrint.document

valADTTypeDef ty generates both the constructor & type parameter for the value ADT corresponding to the type ty

val valueADTDefn : ParserTypes.moduleSig -> PPrint.document

Generates the value ADT definition (enclosed within the ExprToImpl functor)

val interpIsNeeded : ParserTypes.ty -> Base.bool

Given an argument and its type, determines if we need to recursively call interp on the argument for the inner pattern match in interp

val map2 : f:( 'a -> 'b ) -> ('a * 'a) -> 'b * 'b

Applies a function pointwise on a pair

val map3 : f:( 'a -> 'b ) -> ('a * 'a * 'a) -> 'b * 'b * 'b

Applies a function pointwise on a triple

type argPos =

| Fst

| Snd

Auxiliary data type for indicating the position of a non-expr argument to a function

val argPos_of_sexp : Sexplib0.Sexp.t -> argPos

val sexp_of_argPos : argPos -> Sexplib0.Sexp.t

val interpOnce : ParserTypes.ty -> ?nonExprArg:(string * argPos) option -> PPrint.document -> ParserTypes.ident -> ParserTypes.ty -> PPrint.document

Pattern matches interp on one argument of type expr If nonExprArg is Some of some value, it is placed in the appropriate argument position during function application

val interpTwice : ParserTypes.ty -> ParserTypes.ty -> PPrint.document -> ParserTypes.ident -> ParserTypes.ident -> ParserTypes.ty -> PPrint.document

Pattern matches interp on two arguments, both of type expr

val interpExprPatternMatch : (ParserTypes.valDecl * Base.string Base.list) -> PPrint.document

Produces the inner pattern match (interp e) in the interp function

val interpDefn : ParserTypes.moduleSig -> PPrint.document

Generates the definition of the interp function which evaluates exprs

val functorDef : ParserTypes.moduleSig -> sigName:Base.string -> functorName:Base.string -> PPrint.document

Generates the definition of the ExprToImpl functor

val argGen : Base.string -> ParserTypes.ty -> PPrint.document

Produces the code for a monadic bind of arg to a QuickCheck generator producing a value of type ty, where ty must be a non-arrow type (helper function called by genExprPatternRHS)

val genExprPatternRHS : (Base.string * ParserTypes.ty * Base.string Base.list * Base.string * PPrint.document) -> Base.string * (ParserTypes.ty * PPrint.document * Base.string)

Takes in the following arguments: tyConstr: constructor for the ty ADT representing the return type of a function funcTy: function type args: arguments passed to the Expr constructor constr: constructor for the Expr ADT funcApp: PPrint document containing the application of constr onto args

Produces the RHS of the pattern matches in gen_expr, returning a tuple of the form (tyConstr, (ty, patternMatchRHS, nameOfPatternMatch))

Eg. if tyConstr = Bool, ty = Func2(Alpha, Expr, Bool), constr = Mem and funcApp = Mem(x, e), then genExprPatternRHS (tyConstr, ty, args, constr, funcApp) produces the code "let%bind arg = G.int in G.return @@ Mem(x, e)", where G = Base_quickcheck.Generator.

val getExprConstructorWithArgs : Base.string -> ParserTypes.ty -> Base.string -> Base.string * ParserTypes.ty * Base.string Base.list * Base.string * PPrint.document

Takes in tyConstr (constructor for the ty ADT representing the return type of a function), ty (the type of the Expr construcctor), and constr (the string representation of an Expr constructor), and returns a 5-tuple of the form (tyConstr, ty, constructorArgs, constructor, constructor applied to args), eg. (Bool, Func2(Alpha, Expr, Bool), ["x", "e"], "Mem", !^ "Mem(x,e)")

val curry3 : ( ('a * 'b * 'c) -> 'd ) -> 'a -> 'b -> 'c -> 'd

Conversion between the curried/uncurried versions of an arity-3 function

val uncurry3 : ( 'a -> 'b -> 'c -> 'd ) -> ('a * 'b * 'c) -> 'd

val genExprPatterns : ParserTypes.moduleSig -> (Base.string * (ParserTypes.ty * PPrint.document * Base.string)) Base.List.t

Returns an association list of constructors for the ty ADT where each element is the form (<constructor for the ty ADT>, ty)

Generates the definition of the gen_expr Quickcheck generator for the expr datatype

val genExprDef : ParserTypes.moduleSig -> PPrint.document

val functorApp : functorName:ParserTypes.moduleName -> Base.string -> PPrint.document

Produces a PPrint document of an OCaml functor application, where functorName is the name of the functor, and arg is the name of the argument to the functor

val implModuleBindings : functorName:ParserTypes.moduleName -> Base.string -> Base.string -> PPrint.document

Produces the module bindings for the two module implemntations I1, I2, where functorName is the name of the functor that procduces the PBT test harness, and modName1 & modName2 are the names of the two module implementaitons

Code generators for compare_impls.ml, the executable for comparing 2 modules

val displayErrorDef : PPrint.document

Generates the definition of a displayError helper function for displaying error messages when QuickCheck tests fail

val executableImports : pbtFilePath:Base.string -> execFilePath:Base.string -> PPrint.document

Generate requisite imports for the executable file, where filepath is the filepath to the executable file

val obsEquiv : (Base.string * Base.string) -> PPrint.document

Produces Quickcheck code in the executable that tests two modules for observational equivalence based on exprs that return some tyConstr of type ty, and pattern matching on the equivalent valConstrs that they return

val compareImpls : ParserTypes.moduleSig -> PPrint.document

Generate the executable code for testing observational equivalence of two modules